+++ src/Console/Application.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Povils\PHPMND\Console;
 
+use Composer\InstalledVersions;
+use OutOfBoundsException;
 use Povils\PHPMND\Command\RunCommand;
 use Povils\PHPMND\Container;
 use Symfony\Component\Console\Application as BaseApplication;
@@ -14,14 +16,15 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class Application extends BaseApplication
 {
-    public const VERSION = '3.5.0';
+    public const PACKAGE_NAME = 'povils/phpmnd';
+
     private const NAME = 'phpmnd';
 
     private Container $container;
 
     public function __construct(Container $container)
     {
-        parent::__construct(self::NAME, self::VERSION);
+        parent::__construct(self::NAME, self::getPrettyVersion());
 
         $this->setDefaultCommand('run', true);
 
@@ -61,5 +64,24 @@ class Application extends BaseApplication
     protected function getDefaultCommands(): array
     {
         return [new HelpCommand(), new RunCommand()];
+    }
+
+    public static function getPrettyVersion(): string
+    {
+        // Pre 2.0 Composer runtime didn't have this class.
+        if (!class_exists(InstalledVersions::class)) {
+            return 'unknown';
+        }
+
+        try {
+            return (string) InstalledVersions::getPrettyVersion(self::PACKAGE_NAME);
+        } catch (OutOfBoundsException $e) {
+            if (preg_match('#package .*' . preg_quote(self::PACKAGE_NAME, '#') . '.* not installed#i', $e->getMessage()) === 0) {
+                throw $e;
+            }
+
+            // We have a bogus exception: how can PHPMND be not installed if we're here?
+            return 'not-installed';
+        }
     }
 }